Moved the `FundingLink` tests into their own class.
