+++ Tests/AppTests/GithubTests.swift
@@ -397,21 +397,4 @@ class GithubTests: AppTestCase {
         XCTAssertEqual(res, nil)
     }
 
-    func test_fundingLink_missingSchemeFix() async throws {
-        // URL with both a scheme and a host.
-        let ghFundingLink1 = Github.Metadata.FundingLinkNode(platform: .customUrl, url: "https://example.com")
-        let dbFundingLink1 = try XCTUnwrap(FundingLink(from: ghFundingLink1))
-        XCTAssertEqual(dbFundingLink1.url, "https://example.com")
-
-        // URL with a host but no scheme.
-        let ghFundingLink2 = Github.Metadata.FundingLinkNode(platform: .customUrl, url: "example.com")
-        let dbFundingLink2 = try XCTUnwrap(FundingLink(from: ghFundingLink2))
-        XCTAssertEqual(dbFundingLink2.url, "https://example.com")
-
-        // URL with neither.
-        let ghFundingLink3 = Github.Metadata.FundingLinkNode(platform: .customUrl, url: "!@£$%")
-        let dbFundingLink3 = FundingLink(from: ghFundingLink3)
-        XCTAssertNil(dbFundingLink3)
-    }
-
 }
